Supporting firefox
I was using FF 61.0.1 to generate an excel file with the '.xlsx' extension. When trying to open it directly FF wants to change the extension back to 'xls'. At the end I got 'file.xlsx-1.xls' instead of 'file.xlsx' which pops up a warning in Excel (2013). This fix just extends the mime type to support '.xlsx' and so allows FF to open it without warning.